Move StreamDeck and Action to their own files
